chore: rename components file
