+++ src/components/about/member-section.tsx
@@ -1,5 +1,5 @@
 import { Member } from "@/components/common/Member";
-import { Session } from "@/components/common/Session";
+import { Session } from "@/components/common/Section";
 import { Title } from "@/components/common/Title";
 
 export const MemberSession = () => {
